(fix) height method
-> not returning correct path count.

(refactor) rebalance() method.

-> calls function orderedArray and changes this.root
to buildTree(newArray). this.root is required to change
the variable that is returned by an object.

(feat) add orderedArray() method as a helper function.

-> it traverses through the tree using inOrderTraversal
and returns a sorted array.

(refactor) word 'this' to reference this.root was
removed apart from rebalance where it needs to re-assign it.